W111CampEquip: restore scoped equip reconstruction pragmas

+++ src/Battle/btlBoss.c
@@ -1876,7 +1876,7 @@ encounter_default:
 
 #pragma opt_loop_invariants off
 // FUN_002fb860 NONMATCHING
-u16 func_002fb860(BtlUnit* unit, u16 index)
+s64 func_002fb860(BtlUnit* unit, u16 index)
 {
     u16 encounterId = btlBossGetEncounterId();
     switch (encounterId)
